installed and configured CSS modules, added stylesheets and directories for existing components, configured gitignore
