--- src/components/App.jsx
@@ -1,19 +1,0 @@
-import React from 'react';
-// import ProductDetail from './ProductDetail.jsx';
-// import QandA from './QuestionsAndAnswers.jsx';
-// import RandOC from './RelatedItemsAndOutfitCreation.jsx';
-// import RatAndRev from './RatingsAndReviews.jsx';
-
-class App extends React.Component {
-  constructor(props) {
-    super(props)
-  }
-
-  render() {
-    return(
-      <h1>Hello from App</h1>
-    )
-  }
-}
-
-export default App;
