+++ src/components/App/App.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+// import ProductDetail from '../ProductDetail/ProductDetail.jsx';
+// import QandA from '../QuestionsAnsAnswers/QuestionsAndAnswers.jsx';
+// import RandOC from '../RelatedItemsAndOutfitCreation/RelatedItemsAndOutfitCreation.jsx';
+// import RatAndRev from '../RatingsAndReviews/RatingsAndReviews.jsx';
+
+class App extends React.Component {
+  constructor(props) {
+    super(props)
+  }
+
+  render() {
+    return(
+      <h1>Hello from App</h1>
+    )
+  }
+}
+
+export default App;
